requirements.txt
fastapi
uvicorn[standard]
playwright
httpx
pydantic
python-dotenv
capsolver
